refactor: split index.js into modular src/ directory
Resolves #3

Extracted 1149-line monolithic index.js into 21 focused modules:

src/config.js          - env vars, paths, constants
src/utils/rateLimit.js - generic in-memory rate limiter
src/services/
  openclaw.js          - findWorkspace, getOpenClawCommand
  context.js           - getActiveContext (session log reader)
  activity.js          - logActivity, file watcher
  monitor.js           - checkSystemStatus, getVersions
  analyzer.js          - runAnalyzer
src/auth/
  login.html           - login page template (was 270-line inline string)
  sessions.js          - token generation, session store, brute-force protection
  middleware.js        - cookie/header/query key auth middleware
  routes.js            - POST /api/auth, /api/logout
src/routes/
  status.js            - GET /api/status
  logs.js              - GET /api/logs
  tokens.js            - GET/POST /api/tokens
  memory.js            - GET /api/memory
  cron.js              - GET /api/cron, POST /api/run/:id
  process.js           - POST /api/kill, /api/gateway/restart
  config.js            - GET /api/config, /api/check_update
  index.js             - barrel mount
src/app.js             - Express app assembly
src/websocket.js       - WebSocket auth + heartbeat

Result: index.js reduced from 1149 to 70 lines (thin orchestrator).
Zero API surface change. All 14 integration tests pass.
